Added option to treat payload as json-encoded strings and decode them before sending + some refactoring
